fix: corrected transactionReason when a previously cancelled subscription was purchased again (#21)
Hello, I discovered that `transactionReason` was incorrectly set to
`RENEW` instead of `PURCHASE`, for iOS purchases that had been
previously cancelled and then purchased again in the
`purchaseUpdatedListener` callback.

This fix restores the previous logic for determining
`transactionReason`, consistent with the behavior in react-native-iap
versions 12.x.x and 13.x.x and real App Store data encoded in
`jsonRepresentation`.

https://github.com/hyochan/react-native-iap/discussions/3056

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Bug Fixes**
* Improved accuracy in detecting subscription renewals, ensuring renewal
events are correctly identified.
* Reduces instances where subscriptions might appear incorrectly as new
purchases or non-renewals.
* Leads to more reliable subscription status, history, and related
notifications for users.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->